% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-aggregate.R
\name{agg_region}
\alias{agg_region}
\title{Build regional averages for selected region, group, and metric.}
\usage{
agg_region(df, reg, grp, measure, func)
}
\arguments{
\item{df}{input summary data in tableau input format}

\item{reg}{region over which to aggregate}

\item{grp}{group over which to aggregate}

\item{measure}{metric to aggregate ("participants", etc.)}

\item{func}{function to use for aggregation ("sum" or "mean")}
}
\description{
To be called from \code{\link{agg_region_all}}
}
\seealso{
Other functions for preparing regional & national metrics: \code{\link{agg_region_all}},
  \code{\link{drop_incomplete_states}},
  \code{\link{too_few_states}}
}
\concept{functions for preparing regional & national metrics}
