% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-aggregate.R
\name{agg_region_all}
\alias{agg_region_all}
\title{Build regional aggregations for selected metric}
\usage{
agg_region_all(df, regs, measure, func, grps = c("all_sports", "hunt",
  "fish"))
}
\arguments{
\item{df}{input summary data in tableau input format}

\item{regs}{regions over which to aggregate}

\item{measure}{metric to aggregate ("participants", etc.)}

\item{func}{function to use for aggregation ("sum" or "mean")}

\item{grps}{permission groups to aggregate over}
}
\description{
Basically a wrapper for \code{\link{agg_region}} that applies to multiple 
regions and groups.
}
\seealso{
Other functions for preparing regional & national metrics: \code{\link{agg_region}},
  \code{\link{drop_incomplete_states}},
  \code{\link{too_few_states}}
}
\concept{functions for preparing regional & national metrics}
