% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part-rate.R
\name{aggregate_pop}
\alias{aggregate_pop}
\title{Aggregate population by segment (for joining with participant summary)}
\usage{
aggregate_pop(pop_seg, seg = "gender", var = "sex")
}
\arguments{
\item{pop_seg}{input population data frame}

\item{seg}{name of segment to be stored in output data frame}

\item{var}{name of input variable to summarize}
}
\description{
Aggregate population by segment (for joining with participant summary)
}
\seealso{
Other functions for estimating participation rate: \code{\link{est_rate}}
}
\concept{functions for estimating participation rate}
