% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashreg.R
\docType{package}
\name{dashreg}
\alias{dashreg}
\alias{dashreg-package}
\title{dashreg: Produce national/regional dashboards}
\description{
Overview of the core functions for each step in dashboard production
}
\section{1. Prepare summaries using license data (sa-states)}{

\itemize{
  \item use \code{\link{run_state}} to produce dashboard metrics for selected
  state (which is partly a wrapper for \code{\link{run_group}})
  \item based on the dashtemplate package functions 
  \code{\link[dashtemplate]{build_history}}, 
  \code{\link[dashtemplate]{calc_metrics}}, 
  and \code{\link[dashtemplate]{format_metrics}}
}
}

\section{2. Prepare summaries provided directly by states}{

\itemize{
  \item use \code{\link{scale_segs}} if provided segments don't scale to the total.
  \item use \code{\link{est_residents}} if resident breakouts weren't provided.
  \item use \code{\link{est_lm}} with \code{\link{predict_lm}} if you need to
  smooth out artifacts (see 2018-q4 for an example using Florida)
}
}

\section{3. Participation rates}{

\itemize{
  \item use \code{\link{aggregate_pop}} to prepare population data for rate calculation.
  \item use \code{\link{est_rate}} to estimate participation rates
}
}

\section{4. Combine states and Compute National/Regional Metrics}{

\itemize{
  \item use \code{\link{agg_region_all}} to build regional aggregations
  \item this calls \code{\link{agg_region}}
}
}

\section{Prepare ACS population data for participation rate estimates}{

\itemize{
  \item use \code{\link{get_pop}} to pull the reference state totals from
  census "popest" data stored in excel files
  \item use \code{\link{extrapolate_yr}} to impute demographic estimates
  based on state-level totals
}
}

