% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-aggregate.R
\name{drop_incomplete_states}
\alias{drop_incomplete_states}
\title{Exclude states with missing years from aggregation}
\usage{
drop_incomplete_states(df, grp, reg, measure)
}
\arguments{
\item{df}{input summary data in tableau input format}

\item{grp}{group over which to aggregate}

\item{reg}{region over which to aggregate}

\item{measure}{metric to aggregate ("participants", etc.)}
}
\description{
To be called from \code{\link{agg_region}}
}
\seealso{
Other functions for preparing regional & national metrics: 
\code{\link{agg_region_all}()},
\code{\link{agg_region}()},
\code{\link{too_few_states}()}
}
\concept{functions for preparing regional & national metrics}
