% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{est_lm}
\alias{est_lm}
\title{Estimate smoothing factor (using linear regression)}
\usage{
est_lm(df_all, yrs, met, grp, cats)
}
\arguments{
\item{df_all}{summary results in data frame}

\item{yrs}{years that will be used for linear trend estimation}

\item{met}{selected metric (from df_all$metric)}

\item{grp}{selected group (from df_all$group)}

\item{cats}{selected category (from df_all$category)}
}
\description{
Developed in Sep 2019 to adjust for data artifact in FL hunting data
}
\seealso{
Other functions to adjust state results: \code{\link{est_residents}},
  \code{\link{predict_lm}}, \code{\link{scale_segs}}
}
\concept{functions to adjust state results}
