% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{est_residents}
\alias{est_residents}
\title{Estimate resident breakouts}
\usage{
est_residents(df)
}
\arguments{
\item{df}{data frame that holds table of summary results}
}
\value{
Returns a data frame with input df stacked with resident summaries
}
\description{
This was needed in Sep 2019 since state's didn't provided resident breakouts
(oversite on Dan's part). It makes a naive assumption that the demographic
distributions for overall match residents. This probably isn't too egregious
since residents usually dominate the total numbers.
}
\seealso{
Other functions to adjust state results: \code{\link{est_lm}},
  \code{\link{predict_lm}}, \code{\link{scale_segs}}
}
\concept{functions to adjust state results}
