% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{extrapolate_yr}
\alias{extrapolate_yr}
\title{Extrapolate (forward) overall state population for 1 missing year}
\usage{
extrapolate_yr(pop, yr)
}
\arguments{
\item{pop}{population data frame}

\item{yr}{year to be estimated}
}
\description{
This function naively assumes percent change is the same as the previous year.
}
\seealso{
Other functions for acs population estimates: \code{\link{extrapolate_yr_seg}},
  \code{\link{get_pop}}
}
\concept{functions for acs population estimates}
