% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{extrapolate_yr_seg}
\alias{extrapolate_yr_seg}
\title{Extrapolate by-segment state population for missing years}
\usage{
extrapolate_yr_seg(pop_seg, pop, yr, direction)
}
\arguments{
\item{pop_seg}{census sex-by-age data frame (by state)}

\item{pop}{census by state population data frame}

\item{yr}{year to be extrapolated}

\item{direction}{either "forward" (future) or "backward"}
}
\description{
Uses most recent year distributions and pegs to reference state totals
for the year estimated.
}
\seealso{
Other functions for acs population estimates: 
\code{\link{extrapolate_yr}()},
\code{\link{get_pop}()}
}
\concept{functions for acs population estimates}
