% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{get_pop}
\alias{get_pop}
\title{Load reference state total population}
\usage{
get_pop(filename, col_names)
}
\arguments{
\item{filename}{path to excel data file}

\item{col_names}{names to use for columns (passed to \code{\link[readxl]{read_excel}})}
}
\description{
https://www.census.gov/data/tables/time-series/demo/popest/2010s-state-total.html
https://www2.census.gov/programs-surveys/popest/tables/2000-2010/intercensal/state/
}
\seealso{
Other functions for acs population estimates: \code{\link{extrapolate_yr_seg}},
  \code{\link{extrapolate_yr}}
}
\concept{functions for acs population estimates}
