% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg-aggregate.R
\docType{data}
\name{region_relate}
\alias{region_relate}
\title{Relation table for states by region}
\format{A data frame with 10 rows and 2 variables:
\describe{
  \item{state}{2-character state abbreviation}
  \item{region}{region to be used in dashboard}
}}
\usage{
region_relate
}
\description{
Relation table for states by region
}
\seealso{
Other sadashreg data: \code{\link{age_map}}
}
\concept{sadashreg data}
\keyword{datasets}
