% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-states.R
\name{run_group}
\alias{run_group}
\title{Run dashboard summaries for a given permission-timeframe}
\usage{
run_group(cust, lic, sale, yrs, timeframe, group, lic_types)
}
\arguments{
\item{cust}{standardized license data frames}

\item{lic}{standardized license data frames}

\item{sale}{standardized license data frames}

\item{yrs}{years to include in dashboard}

\item{timeframe}{time period covered ("full-year" or "mid-year")}

\item{group}{name of permission group ("fish", "hunt", "all_sports")}

\item{lic_types}{license types (lic$type) included. This should correspond to
a permission group (e.g., c("hunt", "combo") for "hunt" permission).}
}
\description{
Run dashboard summaries for a given permission-timeframe
}
\seealso{
Other functions for SA state dashboard summaries: \code{\link{run_state}}
}
\concept{functions for SA state dashboard summaries}
