% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-states.R
\name{run_state}
\alias{run_state}
\title{Run 3 permission summaries for select state in select time period}
\usage{
run_state(st, yrs, timeframe, outdir, groups = c("all_sports", "hunt",
  "fish"), output_csv = TRUE,
  dir_production = "E:/SA/Data-production/Data-Dashboards")
}
\arguments{
\item{st}{2-character state abbreviation}

\item{yrs}{years to include in dashboard}

\item{timeframe}{time period covered ("full-year" or "mid-year")}

\item{outdir}{location for output csv (named based on input state).}

\item{groups}{names of permissions to be summarized}

\item{output_csv}{if TRUE, will write a csv to outdir}

\item{dir_production}{folder that holds production data}
}
\description{
This basically wraps the entire dashtemplate workflow into 1 function. It
optionally writes a csv output based on the outdir argument.
}
\seealso{
Other functions for SA state dashboard summaries: \code{\link{run_group}}
}
\concept{functions for SA state dashboard summaries}
