% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{scale_segs}
\alias{scale_segs}
\title{Peg segment breakouts to total}
\usage{
scale_segs(df)
}
\arguments{
\item{df}{data frame that holds table of summary results}
}
\description{
This is necessary when state-supplied segment breakouts weren't scaled to
match totals.
}
\seealso{
Other functions to adjust state results: 
\code{\link{est_lm}()},
\code{\link{est_residents}()},
\code{\link{predict_lm}()}
}
\concept{functions to adjust state results}
