% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part-rate.R
\name{aggregate_pop}
\alias{aggregate_pop}
\title{Aggregate population by segment}
\usage{
aggregate_pop(pop_seg, seg = "gender", var = "sex")
}
\arguments{
\item{pop_seg}{input population data frame}

\item{seg}{name of segment to be stored in output data frame}

\item{var}{name of input variable to summarize}
}
\description{
This is a convenience function to prepare the population data to match
the dimension and naming convention needed for joining with participant
summaries produced by \code{\link[salic]{est_part}}.
}
\seealso{
Other functions for estimating participation rate: 
\code{\link{est_rate}()}
}
\concept{functions for estimating participation rate}
