% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part-rate.R
\name{est_rate}
\alias{est_rate}
\title{Add participation rate to summary table for each state}
\usage{
est_rate(dashboard, pop)
}
\arguments{
\item{dashboard}{tableau formatted dashboard data}

\item{pop}{population data prepared with aggregate_pop()}
}
\description{
This takes the ratio of the "residents" metric and the population value and
combines the result with the input dashboard data frame.
}
\seealso{
Other functions for estimating participation rate: \code{\link{aggregate_pop}}
}
\concept{functions for estimating participation rate}
