% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-states.R
\name{run_group}
\alias{run_group}
\title{Run dashboard summaries for a given permission-timeframe}
\usage{
run_group(cust, lic, sale, yrs, timeframe, group, lic_types,
  tests = c(tot = 20, res = 35, sex = 35, agecat = 35),
  scaleup_test = 30)
}
\arguments{
\item{cust}{standardized license data frames}

\item{lic}{standardized license data frames}

\item{sale}{standardized license data frames}

\item{yrs}{years to include in dashboard}

\item{timeframe}{time period covered ("full-year" or "mid-year")}

\item{group}{name of permission group ("fish", "hunt", "all_sports")}

\item{lic_types}{license types (lic$type) included. This should correspond to
a permission group (e.g., c("hunt", "combo") for "hunt" permission).}

\item{tests}{test thresholds for salic::est_part(), salic::est_churn(), and
salic::est_recruit()}

\item{scaleup_test}{test_threshold for salic::scaleup_part()}
}
\description{
Run dashboard summaries for a given permission-timeframe
}
\seealso{
dashtemplate functions: 
\code{\link[dashtemplate]{build_history}} 
\code{\link[dashtemplate]{calc_metrics}}
\code{\link[dashtemplate]{format_metrics}}

Other functions for SA state dashboard summaries: \code{\link{run_state}}
}
\concept{functions for SA state dashboard summaries}
