% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa-states.R
\name{run_state}
\alias{run_state}
\title{Run 3 permission summaries for select state in select time period}
\usage{
run_state(st, yrs, timeframe, outdir, groups = c("all_sports", "hunt",
  "fish"), output_csv = TRUE,
  db_license = file.path("E:/SA/Data-production/Data-Dashboards", st,
  "license.sqlite3"), ...)
}
\arguments{
\item{st}{2-character state abbreviation}

\item{yrs}{years to include in dashboard}

\item{timeframe}{time period covered ("full-year" or "mid-year")}

\item{outdir}{location for output csv (named based on input state).}

\item{groups}{names of permissions to be summarized}

\item{output_csv}{if TRUE, will write a csv to outdir}

\item{db_license}{file path to production data}

\item{...}{additional arguments passed to \code{\link{run_group}}}
}
\description{
This basically wraps the entire dashtemplate workflow into 1 function by (1)
loading sqlite data, (2) applying some minor recoding and data checks, and (3)
Applying \code{\link{run_group}} across the provided groups. 
It returns a summary dataset (formatted for Tableau) and optionally writes 
a csv output based on the outdir argument. The state-level csv files will
ultimately be combined into 1 national/regional file.
}
\seealso{
Other functions for SA state dashboard summaries: \code{\link{run_group}}
}
\concept{functions for SA state dashboard summaries}
